Fixed puppet-chgeck on 101 - 2.0
At the time of submitting this MR the lab doesn't correctly check if lab2.0 has been done correctly (tried to finish with Linux installed and windows not and got no errors).
I'm not sure if the current check-puppet on GitHub has been pushed to Instruqt. If yes, then the current check is not correct and this should fix it.
Also are added different messages according to what agent is missing. 
(Couldn't test the script on a real lab node)